+++ packages/Bio/CHANGELOG.md
@@ -21,6 +21,7 @@
 * WebLogo: Fix horizontal alignment to left while fixWidth
 * WebLogo: Fix layout for fixWidth, fitArea, and normal modes
 * VdRegionsViewer: Fix postponed rendering for tests
+* MacromoleculeDifferenceCellRenderer: Fix to do not use UnitsHandler
 
 ## 2.8.2 (2023-08-01)